move pending withdrawals indicator to a separate file
