+++ packages/arb-token-bridge-ui/src/components/MainContent/MainContent.tsx
@@ -1,8 +1,6 @@
 import { useEffect, useMemo } from 'react'
 import { useMedia, usePrevious } from 'react-use'
 import { motion, AnimatePresence } from 'framer-motion'
-import { twMerge } from 'tailwind-merge'
-import Loader from 'react-loader-spinner'
 
 import { PendingWithdrawalsLoadedState } from '../../util'
 import { useAppState } from '../../state'
@@ -15,6 +13,7 @@ import { TransferPanel } from '../TransferPanel/TransferPanel'
 import { ExploreArbitrum } from './ExploreArbitrum'
 import { HeaderMobileNotification } from '../common/Header'
 import { HeaderAccountPopoverNotification } from '../common/HeaderAccountPopover'
+import { PendingWithdrawalsIndicator } from '../common/PendingWithdrawalsIndicator'
 
 const motionDivProps = {
   layout: true,
@@ -70,54 +69,6 @@ function dedupeWithdrawals(transactions: MergedTransaction[]) {
   })
 
   return Object.values(map)
-}
-
-type PendingWithdrawalsIndicatorProps = {
-  loaderProps?: { width: number; height: number }
-  className?: string
-}
-
-function PendingWithdrawalsIndicator({
-  loaderProps = { width: 14, height: 14 },
-  className = ''
-}: PendingWithdrawalsIndicatorProps) {
-  const { app } = useAppState()
-  const { pwLoadedState, withdrawalsTransformed } = app
-
-  const amount = useMemo(
-    () =>
-      withdrawalsTransformed.filter(
-        tx => tx.status === 'Unconfirmed' || tx.status === 'Confirmed'
-      ).length,
-    [withdrawalsTransformed]
-  )
-
-  if (pwLoadedState === PendingWithdrawalsLoadedState.READY && amount === 0) {
-    return null
-  }
-
-  const bgClassName =
-    pwLoadedState === PendingWithdrawalsLoadedState.LOADING
-      ? 'bg-white'
-      : 'bg-brick'
-
-  return (
-    <div
-      className={twMerge(
-        'flex items-center justify-center rounded-full border-white transition-colors',
-        bgClassName,
-        className
-      )}
-    >
-      {pwLoadedState === PendingWithdrawalsLoadedState.LOADING && (
-        <Loader type="TailSpin" color="black" {...loaderProps} />
-      )}
-
-      {pwLoadedState === PendingWithdrawalsLoadedState.READY && (
-        <span>{amount}</span>
-      )}
-    </div>
-  )
 }
 
 export function MainContent() {
